Added correct library for MPU6050 from github from jazberski.
This one is at least sending data from accelerometer and gyroscope. From start need to wait approximately 3s for sensor to calibrate then it starts send samples. If not in rest during calibration, it's sending nothing.
I'm commiting this with library and pallete build spec to have it saved in repositary.